+++ NEWS.md
@@ -6,6 +6,7 @@ New
 Improved
 - GSEA functions: improved name formatting for enrichment plot
 - `plot_PCA`: fixed subtitle for non-annotated plots (thanks Duncan :) )
+- `plot_GSEA_barplot`: formatting flexibility improvements
 
 Vignettes